% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_MAF.R
\name{annotate_MAF}
\alias{annotate_MAF}
\title{Annotate MAF}
\usage{
annotate_MAF(sample, MAF_database, genome = c("hg19", "hg38"))
}
\arguments{
\item{sample}{\code{VRanges} object of the varscan pileup2cns output}

\item{MAF_database}{\code{GScores} object with MAFs for each position}

\item{genome}{Genome of input object, accepts either hg19 or hg38}
}
\value{
This function returns the \code{VRanges} object with the additional metadata column \code{MAF}.
}
\description{
Annotate VRanges object with population minor allele frequency from gnomAD
}
\examples{
\dontrun{
# Sample with hg19 reference, annotated with gnomAD exome hs37d5
variants <- load_as_VRanges(sample_name = "pt123",
    sample_path = "./patient_123_pileup2cns", genome = "hg19", metadata = TRUE)
library(MafDb.gnomADex.r2.1.hs37d5)
variants <- annotate_MAF(sample = variants,
    MAF_database = MafDb.gnomADex.r2.1.hs37d5, genome = "hg19")
}
\dontrun{
# Sample with hg38 reference, annotated with gnomAD genome GRCh38
variants <- load_as_VRanges(sample_name = "pt321",
    sample_path = "./patient_321_pileup2cns", genome = "hg38", metadata = TRUE)
library(MafDb.gnomAD.r2.0.1.GRCh38)
variants <- annotate_MAF(sample = variants,
    MAF_database = MafDb.gnomAD.r2.0.1.GRCh38, genome = "hg38")
}
}
