% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate Variants}
\usage{
annotate_variants(variant_calls)
}
\arguments{
\item{variant_calls}{\code{VRanges} object from call_all_variants output}
}
\value{
This function returns a \code{DataFrame} object with annotations for each mutation
}
\description{
Wrapper function for oncotate from maftools package to facilitate compatibility with VRanges.
Returns a dataframe with mutations annotated by Oncotator - requires internet connection. Only compatible with hg19 genome.
}
\examples{
\dontrun{
annotated_variants <- variant_calls \%>\% annotate_variants()
}
}
