% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_all_variants.R
\name{call_all_variants}
\alias{call_all_variants}
\title{Call All Variants}
\usage{
call_all_variants(sample, samp_models)
}
\arguments{
\item{sample}{\code{VRanges} object of the varscan pileup2cns output annotated with variant context}

\item{samp_models}{\code{Data.Frame} of context specific error models generated from \code{generate_all_models}}
}
\value{
This function returns a \code{VRanges} with the following metadata:
\itemize{
\item FlankingSeqGroup
	\item Model Used
\item Model pvalue
}
}
\description{
Call all variants from the varscan pileup2cns output and the context-specific error models
In the rare case that no model exists for that context (not enough alternate alleles), then this function uses the varscan assigned pvalue
}
\examples{
\dontrun{
# Get flagged alleles and cosmic mutations
heme_COSMIC <- load_cosmic_mutations(cosmic_mutations_path = "./heme_COSMIC.csv")
flagged_alleles <- get_flagged_alleles(all_sample_names, all_sample_paths,
    exclude_cosmic_mutations = TRUE,cosmic_mutations = heme_COSMIC, cosmic_mut_frequency = 3)

# Load and annotate sample
sample <- load_as_VRanges(sample_name = "pt123",
    sample_path = "./patient_123_pileup2cns", genome = "hg19", metadata = TRUE)
sample <- sequence_context(sample)
library(MafDb.gnomADex.r2.1.hs37d5)
annotated_samp <- annotate_MAF(varscan_output = variants,
    MAF_database = MafDb.gnomADex.r2.1.hs37d5, genome = "hg19")

# Filter model input
samp_model_input <- filter_model_input(model_input = annotated_samp,
    flagged_alleles = flagged_alleles, filter_cosmic_mutations = TRUE,
    cosmic_mutations = heme_COSMIC, cosmic_mut_frequency = 10)

# Generate the error models for this sample
samp_models <- generate_all_models(sample = samp_model_input, plots = FALSE)

# Call variants for the sample based on the previously generated error models
variant_calls <- call_all_variants(annotated_samp, samp_models)
}
}
