% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_variants.R
\name{call_variants}
\alias{call_variants}
\title{Call Variants}
\usage{
call_variants(i, data, flanking_seqs, error_models)
}
\arguments{
\item{i}{Index of the FlankingSeqGroup for which the model is being generated}

\item{data}{\code{Data.Frame} from \code{generate_all_models} with FlankingSeqGroup and three empty columns for parameters of model fit}

\item{flanking_seqs}{String with one of 192 trinucleotide contexts in which variants may be found}

\item{error_models}{\code{Data.Frame} of context specific error models generated from \code{generate_all_models}}
}
\value{
This function returns a \code{VRanges} with the following metadata:
\itemize{
\item FlankingSeqGroup
	\item Model Used
\item Model pvalue
}
}
\description{
Call all variants from the varscan pileup2cns output given a specifci trinucleotide context and the corresponding error model.
}
