% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_pvalues.R
\name{correct_pvalues}
\alias{correct_pvalues}
\title{p-value correction}
\usage{
correct_pvalues(variant_calls, method = "fdr")
}
\arguments{
\item{variant_calls}{\code{VRangesList} or \code{VRanges} object from call_all_variants output}

\item{method}{pvalue correction method to pass onto p.adjust}
}
\value{
This function returns a \code{VRangesList} or \code{VRanges} object with a new metadata column: corrected_pvalue
}
\description{
Apply multiple testing correction (FDR or Bonferroni) to model p-values for each position.
If input is VRangesList, then pvalue correction will be done on a sample-by-sample basis. If input is VRanges, then pvalue correction will be done across all samples.
For mutation calling in a narrow genomic range of specified hotspots, we recommend correcting across all samples.
For mutation calling across all genomic positions sequenced, we recommend applying correction by sample.
}
\examples{
\dontrun{
variants_corrected <- correct_pvalue(variant_calls = variants, method = 'fdr')
}
}
