% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MAPQ.R
\name{filter_MAPQ}
\alias{filter_MAPQ}
\title{Filter MAPQ}
\usage{
filter_MAPQ(sample, MAPQ_cutoff = 59)
}
\arguments{
\item{sample}{\code{VRanges} object of the varscan pileup2cns output}

\item{MAPQ_cutoff}{Minimum acceptable MAPQ score. Default is 59}
}
\value{
This function returns a \code{VRanges} object filtered by MAPQ cutoff
}
\description{
Filter variants to keep those that are equal to or greater than a MAPQ cutoff.
}
\examples{
\dontrun{
variants <- filter_MAPQ(sample = variants, MAPQ_cutoff = 59)
}
}
