% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_alleles.R
\name{flag_alleles}
\alias{flag_alleles}
\title{Flag Alleles}
\usage{
flag_alleles(variants, metadata = FALSE)
}
\arguments{
\item{variants}{\code{VRanges} object from get_flagged_alleles}

\item{metadata}{Logical. Determine whether or not to keep the metadata (annotations) when returning flagged alleles}
}
\value{
This function returns a \code{VRanges} object with the following information:
\itemize{
\item seqnames
	\item ranges
\item ref
\item alt
\item refdepth
\item altdepth
\item sampleNames
\item metadata (optional)
}
}
\description{
Flag alleles that are present in too many samples at high variant allele frequencies as potential errors.
}
