% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate_model}
\alias{generate_model}
\title{Generate Model}
\usage{
generate_model(i, data, groups, model = "auto")
}
\arguments{
\item{i}{Index of the FlankingSeqGroup for which the model is being generated}

\item{data}{\code{Dataframe} from generate_all_models with FlankingSeqGroup and three empty columns for parameters of model fit}

\item{groups}{\code{Dataframe} with 192 rows and 4 columns. First column contains the 192 trinucleotide contexts/FlankingSeqGroup}

\item{model}{Specifying which error model (exp or weibull) to fit. Default is "auto".}
}
\value{
This function returns a \code{dataframe} with the following information:
\itemize{
\item FlankingSeqGroup
	\item Model Used (Exp or Weibull)
\item Model parameter 1
\item Model parameter 2
}
}
\description{
Generate all error models for each trinucleotide variant context. Fits error distribution with either an Exp or Weibull distribution depending on the overall distribution of non-reference alleles.
If the most frequent non-reference allele count is 1 (typically at <10,000 sequencing depth) an Exponential distribution will be fitted and if it is greater than 1 (seen at ultra-deep read depths) a Weibull distribution will be fitted.
}
