% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flagged_alleles.R
\name{get_flagged_alleles}
\alias{get_flagged_alleles}
\title{Get Flagged Alleles}
\usage{
get_flagged_alleles(
  sample_names,
  sample_paths,
  recurrent_mutations = NA,
  memory_saving = FALSE,
  exclude_cosmic_mutations = FALSE,
  cosmic_mutations,
  cosmic_mut_frequency = 3
)
}
\arguments{
\item{sample_names}{Character vector with the names of the samples}

\item{sample_paths}{Character vector with the paths of the samples}

\item{recurrent_mutations}{\code{VRanges} object with chr, pos ref, alt of frequently mutated alleles to remove from model input. \code{GRanges} objects are also accepted, in which case filtering will occur by position.}

\item{memory_saving}{Logical. Option to save memory if you have a lot of samples (e.g. >500 with a 16Gb RAM machine), but takes twice as long}

\item{exclude_cosmic_mutations}{Logical indicating whether or not to exclude cosmic mutations from flagged SNPs}

\item{cosmic_mutations}{\code{VRanges} object with position and substitution of excluded cosmic mutations}

\item{cosmic_mut_frequency}{Mutations with this frequency or above in the cosmic database will be excluded}
}
\value{
This function returns a \code{VRanges} object with the following information:
\itemize{
\item seqnames
	\item ranges
\item ref
\item alt
\item refdepth
\item altdepth
\item sampleNames
\item metadata (optional)
}
}
\description{
Flag alleles that are present in too many samples at high variant allele frequencies as potential errors.
}
\examples{
\dontrun{
# get list of file names
file_names <- list.files(path = "./data/", pattern = "sample")
heme_COSMIC <- load_cosmic_mutations(cosmic_mutations_path = "./heme_COSMIC.csv")

# sample names are first 10 characters of file name
all_sample_names <- substr(file_names, 1, 10)

# file paths are dir/file_name
all_sample_paths <- paste0("./data/", file_names)

# get flagged alleles
flagged_alleles <- get_flagged_alleles(all_sample_names, all_sample_paths,
    exclude_cosmic_mutations = TRUE, cosmic_mutations = heme_COSMIC,
    cosmic_mut_frequency = 3, memory_saving = FALSE)
}
}
