% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemCOSMIC_DC.R
\docType{data}
\name{hemCOSMIC_DC}
\alias{hemCOSMIC_DC}
\title{COSMIC hematological variants}
\format{\code{data.frame} of 754 rows and 46 columns}
\source{
Downloaded from COSMIC https://cancer.sanger.ac.uk/cosmic/download
}
\usage{
hemCOSMIC_DC
}
\description{
Hematological variants from COSMIC80 with annotations and observed frequency
}
\keyword{datasets}
