% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_VRanges.R
\name{intersect_VRanges}
\alias{intersect_VRanges}
\title{Intersect VRanges objects}
\usage{
intersect_VRanges(A, B)
}
\arguments{
\item{A}{\code{VRanges} object to filter from and return}

\item{B}{\code{VRanges} or \code{GRanges} object with alleles to keep}
}
\value{
This function returns \code{VRanges} object A retaining only the alleles that overlap with \code{VRanges} object B.
}
\description{
Identifies features from B that overlap with A, removes the overlapping features from A and returns the remaining portion of A. If both A and B are VRanges objects, then this function is allele specific.
}
\examples{
\dontrun{
variants <- intersect_VRanges(A = variants, B = positions_of_interest)
}
}
