% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_as_VRanges.R
\name{load_as_VRanges}
\alias{load_as_VRanges}
\title{Load as VRanges}
\usage{
load_as_VRanges(sample_name, sample_path, genome, metadata = TRUE)
}
\arguments{
\item{sample_name}{Name of the sample}

\item{sample_path}{Sample file location}

\item{genome}{Reference genome to use.}

\item{metadata}{Logical. Whether to include metadata (VAF, quality scores, strand-specific counts)}
}
\value{
This function returns a \code{VRanges} object with the following information:
\itemize{
\item seqnames
	\item ranges
\item ref
\item alt
\item refdepth
\item altdepth
\item sampleNames
\item metadata (optional)
}
}
\description{
Load varscan pileup2cns output from a given sample as a \code{VRanges} object
}
\examples{
\dontrun{
variants <- load_as_VRanges(sample_name = "pt123", sample_path = "./patient_123_pileup2cns",
genome = "hg19", metadata = TRUE)
}
}
