% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cosmic_mutations.R
\name{load_cosmic_mutations}
\alias{load_cosmic_mutations}
\title{Load Cosmic Mutations}
\usage{
load_cosmic_mutations(cosmic_mutations_path = "hemCOSMIC.csv")
}
\arguments{
\item{cosmic_mutations_path}{Path to csv file containing COSMIC mutations}
}
\value{
This function returns a \code{VRanges} object including:
\itemize{
\item seqnames
	\item ranges
\item ref
\item alt
\item metadata (including observed frequency)
}
}
\description{
Loads in Cosmic mutations from file and present it as a VRanges object.
}
\examples{
\dontrun{
heme_COSMIC <- load_cosmic_mutations(cosmic_mutations_path = "heme_COSMIC.csv")
}
}
