% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recurrent_mutations.R
\name{load_recurrent_mutations}
\alias{load_recurrent_mutations}
\title{Load File with Recurrent Mutations to exclude}
\usage{
load_recurrent_mutations(file, genome = c("hg19", "hg38"))
}
\arguments{
\item{file}{Location of file with recurrent mutations}

\item{genome}{Reference genome to use, default is hg19}
}
\value{
This function returns a \code{VRanges} object including:
\itemize{
\item seqnames
	\item ranges
	\item ref
	\item alt
}
}
\description{
Loads in recurrent mutations from tab or comma-delimited file with four columns (Chr, Pos, Ref, Alt) and no header, presenting it as a VRanges file.
}
\examples{
\dontrun{
mutations_to_exclude <- load_recurrent_mutations(file = "COSMIC_Heme_mutations_freq10.txt", genome = "hg19")
}
}
