% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context.R
\name{sequence_context}
\alias{sequence_context}
\title{Sequence Context}
\usage{
sequence_context(sample, genome = c("hg19", "hg38"), context = 3)
}
\arguments{
\item{sample}{\code{VRanges} object of the varscan pileup2cns output}

\item{genome}{Genome of input object, either hg19 or hg38}

\item{context}{Number of nucleotides to return. For example, 3 would return three nucleotides with the position of the variant being the middle nucleotide.}
}
\value{
This function returns the \code{VRanges} object with the additional metadata column \code{context}.
}
\description{
Get sequence context for a \code{VRanges} object of the varscan pileup2cns output
}
\examples{
\dontrun{
variants <- load_as_VRanges(sample_name = "pt123",
    sample_path = "./patient_123_pileup2cns", genome = "hg19", metadata = TRUE)
variants <- sequence_context(sample = variants, context = 3)
}
}
