% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_VRanges.R
\name{subtract_VRanges}
\alias{subtract_VRanges}
\title{Subtract VRanges objects}
\usage{
subtract_VRanges(A, B)
}
\arguments{
\item{A}{\code{VRanges} object to filter from and return}

\item{B}{\code{VRanges} or \code{GRanges} object with alleles to filter out}
}
\value{
This function returns \code{VRanges} object A with alleles from \code{VRanges} object B filtered out
}
\description{
Identifies features from B that overlap with A, removes the overlapping features from A and returns the remaining portion of A. If both A and B are VRanges objects, then this function is allele specific.
}
\examples{
\dontrun{
variants <- subtract_VRanges(A = variants, B = flagged_alleles)
}
}
