% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate Variants}
\usage{
annotate_variants(variant_calls, genome = c("hg19", "hg38"), batchsize = 80)
}
\arguments{
\item{variant_calls}{\code{VRanges} object from call_all_variants output}

\item{genome}{Genome of input object, accepts either hg19 or hg38}

\item{batchsize}{Size of variant batches to send to CellBase - the API will time out if processing >100 variants at a time. If it does time out, reduce the batchsize.}
}
\value{
This function returns a \code{DataFrame} object with annotations for each variant.
}
\description{
Wrapper function for variant annotation from CellBase through cellbaseR
Returns a dataframe with mutations annotated by CellBase - requires internet connection.
}
\examples{
\dontrun{
annotated_variants <- annotate_variants(variant_calls, genome="hg19")
}
}
