% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_pvalues.R
\name{correct_pvalues}
\alias{correct_pvalues}
\title{p-value correction}
\usage{
correct_pvalues(variant_calls, by_sample = FALSE)
}
\arguments{
\item{variant_calls}{\code{VRanges} object from call_all_variants output}

\item{by_sample}{Logical. Option to apply multiple testing correction by sample.}
}
\value{
This function returns a \code{VRanges} object with a new metadata column: corrected_pvalue
}
\description{
Apply bonferroni multiple testing correction to model p-values for each position.
For mutation calling in a narrow genomic range of specified hotspots, we recommend correcting across all samples.
For mutation calling across all genomic positions sequenced, we recommend applying correction by sample.
}
\examples{
\dontrun{
variants_corrected <- correct_pvalue(variant_calls = variants, by_sample = FALSE)
}
}
