% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_model_input.R
\name{filter_model_input}
\alias{filter_model_input}
\title{Filter Model Input}
\usage{
filter_model_input(
  model_input,
  flagged_alleles = NA,
  MAF_cutoff = 0.001,
  VAF_cutoff = 0.05,
  filter_cosmic_mutations = FALSE,
  cosmic_mutations,
  cosmic_mut_frequency = 10,
  MAPQ_cutoff_ref = 59,
  MAPQ_cutoff_alt = 59,
  recurrent_mutations = NA
)
}
\arguments{
\item{model_input}{\code{VRanges} object annotated with mutation context and population minor allele frequency}

\item{flagged_alleles}{\code{VRanges} object with high VAF alleles flagged as being present in too many samples}

\item{MAF_cutoff}{Population Minor Allele Frequency cutoff: variants at or above this cutoff are excluded}

\item{VAF_cutoff}{Sample Variant Allele Frequency cutoff: variants at or above this cutoff are excluded}

\item{filter_cosmic_mutations}{Logical indicating whether or not to filter cosmic mutations}

\item{cosmic_mutations}{\code{VRanges} object with position and substitution of excluded cosmic mutations}

\item{cosmic_mut_frequency}{Mutations with this frequency or above in the cosmic database will be excluded}

\item{MAPQ_cutoff_ref}{Minimum acceptable MAPQ score for reference allele}

\item{MAPQ_cutoff_alt}{Minimum acceptable MAPQ score for alternate allele}

\item{recurrent_mutations}{\code{VRanges} object with chr, pos ref, alt of frequently mutated alleles to remove from model input. \code{GRanges} objects are also accepted, in which case filtering will occur by position.}
}
\value{
This function returns a filtered \code{VRanges} object.
}
\description{
Filter variants to remove flagged alleles, polymorphisms, cosmic mutations, and high VAF prior to error model generation.
}
\examples{
\dontrun{
# Get flagged alleles and cosmic mutations
heme_COSMIC <- load_cosmic_mutations(cosmic_mutations_path = "./heme_COSMIC.csv")
flagged_alleles <- get_flagged_alleles(all_sample_names, all_sample_paths,
    exclude_cosmic_mutations = TRUE, cosmic_mutations = heme_COSMIC, cosmic_mut_frequency = 3)

# Load and annotate sample
samp <- load_as_VRanges(sample_name = "pt123",
    sample_path = "./patient_123_pileup2cns", genome = "hg19", metadata = TRUE)
samp <- sequence_context(samp)
library(MafDb.gnomADex.r2.1.hs37d5)
annotated_samp <- annotate_MAF(varscan_output = variants,
    MAF_database = MafDb.gnomADex.r2.1.hs37d5, genome = "hg19")

# Filter model input
samp_model_input <- filter_model_input(model_input = annotated_samp,
    flagged_alleles = flagged_alleles, filter_cosmic_mutations = TRUE,
    cosmic_mutations = heme_COSMIC, cosmic_mut_frequency = 10)
}
}
