% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate_model}
\alias{generate_model}
\title{Generate Model}
\usage{
generate_model(i, data, groups, model = "auto")
}
\arguments{
\item{i}{Index of the FlankingSeqGroup for which the model is being generated}

\item{data}{\code{Dataframe} from generate_all_models with FlankingSeqGroup and three empty columns for parameters of model fit}

\item{groups}{\code{Dataframe} with 192 rows and 4 columns. First column contains the 192 trinucleotide contexts/FlankingSeqGroup}

\item{model}{Specifying which error model (exp or weibull) to fit. Default is "auto".}
}
\value{
This function returns a \code{dataframe} with the following information:
\itemize{
\item FlankingSeqGroup
	\item Model Used (Exp or Weibull)
\item Model parameter 1
\item Model parameter 2
}
}
\description{
Generate error model given a specific trinucleotide context. Fits error distribution with Exp model for <10,000 sequencing depth and Weibull model for >=10,000 depth.
}
