% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flagged_alleles.R
\name{get_flagged_alleles}
\alias{get_flagged_alleles}
\title{Get Flagged Alleles}
\usage{
get_flagged_alleles(
  sample_names,
  sample_paths,
  genome,
  recurrent_mutations = NA,
  memory_saving = FALSE,
  starting_percentile = 99,
  interval = 0.001,
  MAPQcutoff = 59
)
}
\arguments{
\item{sample_names}{Character vector with the names of the samples}

\item{sample_paths}{Character vector with the paths of the samples}

\item{recurrent_mutations}{\code{VRanges} object with chr, pos ref, alt of frequently mutated alleles to remove from model input. \code{GRanges} objects are also accepted, in which case filtering will occur by position.}

\item{memory_saving}{Logical. Option to save memory if you have a lot of samples (e.g. >500 with a 16Gb RAM machine), but takes twice as long}

\item{starting_percentile}{Lower VAF percentile to start looking for alleles to flag. Default is 99, but can use 95 if you want to flag more alleles (more conservative)}

\item{interval}{VAF interval to iterate through for flagging alleles. Default is 0.001}

\item{MAPQcutoff}{Minimum acceptable MAPQ score; positions below this cutoff will be excluded. Default is 59}
}
\value{
This function returns a \code{VRanges} object with the following information:
\itemize{
\item seqnames
	\item ranges
\item ref
\item alt
\item refdepth
\item altdepth
\item sampleNames
\item metadata (optional)
}
}
\description{
Flag alleles that are present in too many samples at high variant allele frequencies as potential errors.
}
\examples{
\dontrun{
# get list of file names
file_names <- list.files(path = "./data/", pattern = "sample")
hemeCOSMIC_3 <- load_recurrent_mutations("example_data/COSMIC_heme_freq3.txt", genome = "hg19")

# sample names are first 10 characters of file name
all_sample_names <- substr(file_names, 1, 10)

# file paths are dir/file_name
all_sample_paths <- paste0("./data/", file_names)

# get flagged alleles
flagged_alleles <- get_flagged_alleles(all_sample_names, all_sample_paths,
    recurrent_mutations = hemeCOSMIC_3, memory_saving = FALSE)
}
}
